fixed truncating conversion from time_t to unsigned int in random module

+++ yosen_lang_std/random_module_init.cpp
@@ -4,7 +4,7 @@
 
 EXTERNC YOSENEXPORT void _ys_init_module()
 {
-	srand(time(NULL));
+	srand((unsigned int)time(NULL));
 
 	auto& env = YosenEnvironment::get();
